This is a simple Registration form.
The form has basic user input like name, Date of birth, address, mobile number and a checkbox to check terms and conditions. After submitting the form all the details of the form is listed on the right side of the form and a dialogue box appears saying "Registration is successful".